Split handlerParam decorators
