+++ src/decorators/handlerParam/handlerParam.ts
@@ -1,0 +1,57 @@
+import { MetaKey } from '../../consts'
+import express from 'express'
+
+const metaKey = MetaKey.handlerParam
+
+export type HandlerParamSelector<T> = (
+  req: express.Request,
+  res: express.Response,
+  next: express.NextFunction
+) => T
+
+export interface HandlerParamMeta<T> {
+  index: number
+  selector: HandlerParamSelector<T>
+}
+
+export type HandlerParamMetaList = HandlerParamMeta<any>[]
+
+export function getHandlerParamMetaList(
+  controller: any,
+  propertyKey: string
+): HandlerParamMetaList {
+  const metaList = Reflect.getMetadata(metaKey, controller, propertyKey)
+  if (metaList == null) return []
+  return metaList
+}
+
+export function setHandlerParamMetaList(
+  controller: any,
+  meta: HandlerParamMetaList,
+  propertyKey: string
+): void {
+  Reflect.defineMetadata(metaKey, meta, controller, propertyKey)
+}
+
+export function handlerParam<T>(selector: HandlerParamSelector<T>) {
+  return function controllerDecorator(
+    target: any,
+    propertyKey: string,
+    index: number
+  ) {
+    let previousHandlerParamList = getHandlerParamMetaList(
+      target.constructor,
+      propertyKey
+    )
+
+    const meta: HandlerParamMetaList = [
+      {
+        index,
+        selector
+      },
+      ...previousHandlerParamList
+    ]
+
+    setHandlerParamMetaList(target.constructor, meta, propertyKey)
+  }
+}
