+++ src/decorators/handlerParam/reqHeaders.ts
@@ -1,0 +1,7 @@
+import { HandlerParamSelector, handlerParam } from './handlerParam'
+
+export function reqHeaders(paramName?: string) {
+  const selector: HandlerParamSelector<any> =
+    paramName == null ? req => req.headers : req => req.headers[paramName]
+  return handlerParam(selector)
+}
