+++ src/decorators/handlerParam/reqSession.ts
@@ -1,0 +1,5 @@
+import { handlerParam } from './handlerParam'
+
+export function reqSession() {
+  return handlerParam(req => req.session)
+}
